Add CHAINER_TEST_PAIRWISE_PARAMETERIZATION and enable it only in Travis CI
