+++ tests/chainer_tests/conftest.py
@@ -1,8 +1,6 @@
 import pytest
 
 from chainer.backends import cuda
-from chainer import testing
-from chainer.testing import parameterized
 import chainerx
 
 
@@ -11,20 +9,6 @@ if not chainerx.is_available():
     # TODO(kmaehashi) This is an tentative fix. This file should be removed
     # once chainer-test supports ChainerX.
     pytest.mark.chainerx = pytest.mark.skip
-
-
-def pytest_collection(session):
-    # Perform pairwise testing.
-    # TODO(kataoka): This is a tentative fix. Discuss its public interface.
-    pairwise_product_dict = parameterized._pairwise_product_dict
-    testing.product_dict = pairwise_product_dict
-    parameterized.product_dict = pairwise_product_dict
-
-
-def pytest_collection_finish(session):
-    product_dict = parameterized._product_dict_orig
-    testing.product_dict = product_dict
-    parameterized.product_dict = product_dict
 
 
 def pytest_runtest_teardown(item, nextitem):
