+++ tests/chainerx_tests/conftest.py
@@ -1,7 +1,5 @@
 import pytest
 
-from chainer import testing
-from chainer.testing import parameterized
 import chainerx.testing
 
 from chainerx_tests import cuda_utils
@@ -9,20 +7,6 @@ from chainerx_tests import cuda_utils
 
 def pytest_configure(config):
     _register_cuda_marker(config)
-
-
-def pytest_collection(session):
-    # Perform pairwise testing.
-    # TODO(kataoka): This is a tentative fix. Discuss its public interface.
-    pairwise_product_dict = parameterized._pairwise_product_dict
-    testing.product_dict = pairwise_product_dict
-    parameterized.product_dict = pairwise_product_dict
-
-
-def pytest_collection_finish(session):
-    product_dict = parameterized._product_dict_orig
-    testing.product_dict = product_dict
-    parameterized.product_dict = product_dict
 
 
 def pytest_runtest_setup(item):
